read Blackboard data correctly from the SFC file, and construct a BlackboardPart and thereby render the text of blackboards

+++ SFCFile.h
@@ -244,7 +244,7 @@ public:
 class SFCBlackboard : public SFCCompoundObject {
 public:
 	uint32 textx, texty;
-	uint16 backgroundcolour, chalkcolour, aliascolour;
+	uint32 backgroundcolour, chalkcolour, aliascolour;
 	std::vector<std::pair<uint32, std::string> > strings;
 
 	SFCBlackboard(SFCFile *p) : SFCCompoundObject(p) { }
